+++ R/wu_gfr_snippet.R
@@ -7,13 +7,12 @@
 #'
 #' @details This generates the same HTML content as \code{\link{wu_gfr_report}},
 #' however only the report elements are saved to the HTML file. In other words
-#' this function doesn't produce a full
+#' this function doesn't produce a full HTML report
 #'
 #'
 #' @seealso \code{\link{wu_gfr_elems}}, \code{\link{wu_gfr_report}}
 #' @importFrom knitr image_uri
 #' @importFrom htmltools tagList img
-#' @importFrom ggplot2 ggsave
 #' @importFrom magrittr extract2
 #'
 #' @export